requirements.txt
https://download.pytorch.org/whl/cpu/torch-1.8.1%2Bcpu-cp38-cp38-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.9.1%2Bcpu-cp38-cp38-linux_x86_64.whl
absl-py==0.12.0
aiohttp==3.7.4.post0
APScheduler==3.7.0
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
Cython==0.29.22
datasets==1.2.1
decorator==4.4.2
demjson==2.2.4
dill==0.3.3
editdistance==0.5.3
fasttext==0.9.1
filelock==3.0.12
Flask==1.1.2
Flask-APScheduler==1.12.2
Flask-SQLAlchemy==2.5.1
fsspec==0.9.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.0
google-auth==1.28.1
google-auth-oauthlib==0.4.4
greenlet==1.0.0
grpcio==1.37.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.10.0
importlib-resources==5.1.2
iopath==0.1.7
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
lime==0.2.0.1
llvmlite==0.36.0
lmdb==0.98
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
multidict==5.1.0
multiprocess==0.70.11.1
networkx==2.5.1
nltk==3.4.5
numba==0.53.1
numpy==1.20.2
oauthlib==3.1.0
omegaconf==2.0.6
onnx=1.9.0=pypi_0
onnxruntime=1.7.0=pypi_0
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
portalocker==2.3.0
prompt-toolkit==3.0.18
protobuf==3.15.8
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycocotools==2.0.2
Pygments==2.8.1
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.2.7
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
regex==2021.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.44
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
sentencepiece==0.1.95
shap==0.39.0
six==1.15.0
sklearn==0.0
slicer==0.0.7
smmap==4.0.0
SQLAlchemy==1.4.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.4.8
tokenizers==0.9.2
torchaudio==0.8.1
torchmetrics==0.2.0
torchray==1.0.0.2
torchtext==0.5.0
tqdm==4.49.0
traitlets==5.0.5
transformers==3.4.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==1.0.1
wincertstore==0.2
xxhash==2.0.0
yarl==1.6.3
zipp==3.4.1
